create signup page, move landing to its own component, move sass around
