+++ src/App.scss
@@ -1,4 +1,6 @@
 @import 'styles/variables';
+@import 'styles/mixins';
+@import 'styles/form.scss';
 
 .clickable {
   opacity: 1;
