+++ src/components/common/Button/styles.scss
@@ -3,13 +3,14 @@
 .button {
   background-color: $color-black;
   border: none;
-  border-radius: 5px;
   color: $color-white;
   font-size: $font-size-button;
   font-weight: bold;
   margin: 0.3rem 0;
   padding: 0.8rem;
   width: 100%;
+  box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.3);
+  transition: 0.2s ease;
 }
 
 .button:hover {
