+++ src/components/common/Smilies/index.js
@@ -1,0 +1,5 @@
+import './styles.scss';
+
+const Smilies = () => <img src="/smilies.png" alt="smilies" className="smilies" />;
+
+export default Smilies;
