+++ src/components/common/Smilies/styles.scss
@@ -1,0 +1,4 @@
+.smilies {
+  max-height: 150px;
+  margin-top: 4rem;
+}
