+++ src/components/form/Select/styles.scss
@@ -1,14 +1,15 @@
 @import 'styles/variables';
 
-.Input {
+.Select {
   margin: 0.3rem 0 0.8rem;
 
-  input {
+  select {
     border: 1px solid $color-black;
     box-sizing: border-box;
     font-size: 16px;
     padding: 0.8rem;
     width: 100%;
+    text-align: center;
     &.error {
       border: 1px solid $color-red;
     }
