+++ src/components/landing/LandingScreen/index.js
@@ -1,0 +1,13 @@
+import VideoContainer from '../VideoContainer';
+import './styles.scss';
+
+const LandingScreen = ({ children }) => {
+  return (
+    <div className="screen-container">
+      <div className="main-container">{children}</div>
+      <VideoContainer />
+    </div>
+  );
+};
+
+export default LandingScreen;
