+++ src/components/landing/LandingScreen/styles.scss
@@ -1,0 +1,16 @@
+@import 'styles/variables';
+@import 'styles/mixins';
+
+.screen-container {
+  @include flex();
+  width: 100vw;
+  height: 100vh;
+
+  .main-container {
+    @include flex(column, center, center);
+    width: 50vw;
+    height: 100vh;
+
+    background: white;
+  }
+}
