+++ src/components/landing/VideoContainer/styles.scss
@@ -20,8 +20,4 @@
       -ms-transform: translate(-50%, -50%);
     }
   }
-
-  .facebook {
-    margin-bottom: 0.5rem;
-  }
 }
