+++ src/components/login/LoginForm/styles.scss
@@ -1,0 +1,4 @@
+.facebook {
+  margin-bottom: 0.5rem;
+  margin-top: 2rem;
+}
