+++ src/components/signup/SignupForm/styles.scss
@@ -1,0 +1,3 @@
+.signup {
+  width: 35%;
+}
