+++ src/pages/Login/styles.scss
@@ -1,68 +1,2 @@
 @import 'styles/variables';
 @import 'styles/mixins';
-
-.screen-container {
-  @include flex();
-  width: 100vw;
-  height: 100vh;
-
-  .login-container {
-    @include flex(column, center, center);
-    width: 50vw;
-    height: 100vh;
-
-    background: white;
-
-    .form-login {
-      @include flex();
-      color: $color-black;
-      max-width: 216.77px;
-    }
-
-    .title {
-      font-family: 'Open Sans', sans-serif;
-      font-size: 20px;
-      line-height: 23px;
-      letter-spacing: 3px;
-      font-weight: bolder;
-      margin: 2rem 0 2rem 0;
-    }
-
-    .subtitle {
-      margin: 0;
-      font-size: 18px;
-      line-height: 32px;
-      font-weight: 500;
-    }
-
-    .description {
-      width: 321px;
-      font-family: 'Open Sans', sans-serif;
-      font-size: 14px;
-      line-height: 26px;
-      text-align: center;
-      margin: 0.5rem 0 3rem 0;
-    }
-  }
-}
-
-.smilies {
-  max-height: 150px;
-  margin-top: 4rem;
-}
-
-.form-label {
-  @include flex();
-  width: 100%;
-}
-
-.separator {
-  width: 75%;
-  border-bottom: 1px solid black;
-  margin-bottom: 1rem;
-}
-
-.link-to {
-  text-decoration: none;
-  color: inherit;
-}
