+++ src/pages/Signup/styles.scss
@@ -1,0 +1,2 @@
+@import 'styles/mixins';
+@import 'styles/variables';
